revert: deleted the whole SignupImage component
So this component was built for upload a profile image into the firebase storage and use it on the app.
And then I realized that this feature (shwoing the photo that we uploaded with the url link that firebase gives us)
is unavailable for my country (Iran).
It's hard to see something that you have nothing to do with, is going to be a barrier to your work.